+++ lib/open_feature/provider.rb
@@ -3,10 +3,10 @@
 
 module OpenFeature
   # Interface that providers must implement.
-  module Provider
+  class Provider
     extend T::Sig
     extend T::Helpers
-    interface!
+    abstract!
 
     sig { abstract.returns(ProviderMetadata) }
     def metadata; end