+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- *Breaking* `OpenFeature::Provider` changed from a module interface to an abstract class to support default method implementations.
+
 ## [0.2.0] - 2023-05-17
 
 ### Added